+++ apps/router-e2e/__e2e__/native-tabs/app/_layout.tsx
@@ -7,7 +7,7 @@ import { Appearance, Platform, useColorScheme } from 'react-native';
 import { ActiveTabsContext } from '../utils/active-tabs-context';
 
 if (process.env.EXPO_OS !== 'web') {
-  Appearance.setColorScheme(null);
+  Appearance.setColorScheme('unspecified');
 }
 
 export default function Layout() {
@@ -47,7 +47,7 @@ export default function Layout() {
             <Icon
               // selectedColor="deepNavy"
               sf="applewatch.side.right"
-              drawable="ic_phone"
+              androidSrc={<VectorIcon family={MIcons} name="watch" />}
             />
           </NativeTabs.Trigger>
           {activeTabs.map((tab, index) => (