+++ packages/@expo/cli/src/start/server/DevToolsPluginManager.ts
@@ -1,18 +1,11 @@
 import type { ModuleDescriptorDevTools } from 'expo-modules-autolinking/exports';
 
+import { DevToolsPlugin } from './DevToolsPlugin';
+import { Log } from '../../log';
+
 const debug = require('debug')('expo:start:server:devtools');
 
 export const DevToolsPluginEndpoint = '/_expo/plugins';
-
-interface AutolinkingPlugin {
-  packageName: string;
-  packageRoot: string;
-  webpageRoot: string;
-}
-
-export interface DevToolsPlugin extends AutolinkingPlugin {
-  webpageEndpoint: string;
-}
 
 export default class DevToolsPluginManager {
   private plugins: DevToolsPlugin[] | null = null;
@@ -20,14 +13,9 @@ export default class DevToolsPluginManager {
   constructor(private projectRoot: string) {}
 
   public async queryPluginsAsync(): Promise<DevToolsPlugin[]> {
-    if (this.plugins) {
-      return this.plugins;
+    if (!this.plugins) {
+      this.plugins = await this.queryAutolinkedPluginsAsync(this.projectRoot);
     }
-    const plugins = (await this.queryAutolinkedPluginsAsync(this.projectRoot)).map((plugin) => ({
-      ...plugin,
-      webpageEndpoint: `${DevToolsPluginEndpoint}/${plugin.packageName}`,
-    }));
-    this.plugins = plugins;
     return this.plugins;
   }
 
@@ -37,7 +25,7 @@ export default class DevToolsPluginManager {
     return plugin?.webpageRoot ?? null;
   }
 
-  private async queryAutolinkedPluginsAsync(projectRoot: string): Promise<AutolinkingPlugin[]> {
+  private async queryAutolinkedPluginsAsync(projectRoot: string): Promise<DevToolsPlugin[]> {
     const autolinking: typeof import('expo/internal/unstable-autolinking-exports') = require('expo/internal/unstable-autolinking-exports');
     const linker = autolinking.makeCachedDependenciesLinker({ projectRoot });
     const revisions = await autolinking.scanExpoModuleResolutionsForPlatform(linker, 'devtools');
@@ -48,6 +36,8 @@ export default class DevToolsPluginManager {
       )
     ).filter((maybePlugin) => maybePlugin != null);
     debug('Found autolinked plugins', plugins);
-    return plugins;
+    return plugins
+      .map((pluginInfo) => new DevToolsPlugin(pluginInfo, this.projectRoot))
+      .filter((p) => p != null) as DevToolsPlugin[];
   }
 }