+++ packages/@expo/cli/src/start/server/metro/createServerComponentsMiddleware.ts
@@ -6,9 +6,9 @@
  */
 import { getMetroServerRoot } from '@expo/config/paths';
 import type { SerialAsset } from '@expo/metro-config/build/serializer/serializerAssets';
-import { getRscMiddleware } from '@expo/server/private';
 import assert from 'assert';
 import type { EntriesDev } from 'expo-router/build/rsc/server';
+import { getRscMiddleware } from 'expo-server/private';
 import path from 'path';
 import url from 'url';