+++ packages/expo-modules-core/common/cpp/fabric/ExpoViewState.h
@@ -28,11 +28,28 @@ public:
       _height = std::numeric_limits<float>::quiet_NaN();
     }
   };
+  
+  static ExpoViewState withStyleDimensions(float styleWidth, float styleHeight) {
+    ExpoViewState state;
+    if (styleWidth >= 0) {
+      state._styleWidth = styleWidth;
+    } else {
+      state._styleWidth = std::numeric_limits<float>::quiet_NaN();
+    }
+    if (styleHeight >= 0) {
+      state._styleHeight = styleHeight;
+    } else {
+      state._styleHeight = std::numeric_limits<float>::quiet_NaN();
+    }
+    return state;
+  }
 
 #ifdef ANDROID
   ExpoViewState(ExpoViewState const &previousState, folly::dynamic data)
   : _width((float)data["width"].getDouble()),
-    _height((float)data["height"].getDouble()){};
+    _height((float)data["height"].getDouble()),
+    _styleWidth(data.count("styleWidth") ? (float)data["styleWidth"].getDouble() : std::numeric_limits<float>::quiet_NaN()),
+    _styleHeight(data.count("styleHeight") ? (float)data["styleHeight"].getDouble() : std::numeric_limits<float>::quiet_NaN()){};
   folly::dynamic getDynamic() const {
     return {};
   };
@@ -44,6 +61,8 @@ public:
   
   float _width = std::numeric_limits<float>::quiet_NaN();
   float _height = std::numeric_limits<float>::quiet_NaN();
+  float _styleWidth = std::numeric_limits<float>::quiet_NaN();
+  float _styleHeight = std::numeric_limits<float>::quiet_NaN();
 
 };